Moved dashboard into the logging package.
