+++ src/main/java/chain/ChainManager.kt
@@ -5,6 +5,7 @@ import communication.Message
 import communication.SyncRequest
 import communication.TransmissionType
 import data.*
+import logging.Dashboard
 import logging.Logger
 import manager.*
 import org.apache.commons.codec.digest.DigestUtils
